% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyLogreg.R
\name{applyLogreg}
\alias{applyLogreg}
\title{Apply logistic regression.}
\usage{
applyLogreg(dataTrain = NULL, dataTest = NULL, frmla = NULL, outcome = "y")
}
\arguments{
\item{dataTrain}{Training subset (80 percent) of the full sample.}

\item{dataTest}{Test subset (20 percent) of the full sample.}

\item{frmla}{An object of the class \code{formula}, used for the logistic regression model.}

\item{outcome}{Character string. Name of the outcome variable, in this study it was 'newAud' (new onset of alcohol use disorder).}
}
\value{
a list with two data.frames as elements (names: ApparentCV, TestCV), each data.frame having three columns:
\enumerate{
\item observed Observed outcome (0 = absent, 1 = present).
\item predicted Model-based probability estimation of the outcome being present.
\item ids Row numbers of the total sample (before data has been split into training and test subsets).
}
}
\description{
Apply the logistic regression model to the data.
}
\examples{
# See the accompanying R script and this package's vignette,
# section Repeated k-fold cross-validation.
}
\author{
Marcel Miché
}
