% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_optimism1.R
\name{boot_optimism1}
\alias{boot_optimism1}
\title{Run bootstrap performance optimism estimation.}
\usage{
boot_optimism1(data, outcome = "newAud", B = 200)
}
\arguments{
\item{data}{Dataset which provides the required columns (see \strong{Details}).}

\item{outcome}{Character string. Name of the outcome variable, in this study it was 'firstAud' (first onset of alcohol use disorder).}

\item{B}{number of bootstrap resamples to run (we set B = 500).}
}
\value{
a data.frame with as many rows as the original data and with as many columns, as the function's argument B was set to, e.g., 500. Each column contains predicted probabilities. They were produced by training the logistic regression model with a bootstrapped version of the original data, which then was cross-validated with the original data. The predicted probabilities stem from this cross-validation with the original data.
}
\description{
Apply the bootstrap evaluation of the model performance.
}
\details{
Dataset must contain the outcome and all eight predictor variables. This function is a modification of the function \code{boot_optimism} of the R package \code{pminternal}.
Marcel Miché modified the original code. Enter predictAUDPsyCoLausDev::boot_optimism1 in the R console to see the raw code, including the code that I outcommented.
}
\examples{
# See the accompanying R script and this package's vignette,
# section 500 bootstrap prediction performance results.
}
\references{
\insertRef{pbapply2023}{predictAUDPsyCoLaus}

\insertRef{R2024}{predictAUDPsyCoLaus}
}
\author{
Stephen Rhodes (see \strong{Details})
}
