% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_optimismDCA.R
\name{boot_optimismDCA}
\alias{boot_optimismDCA}
\title{Run bootstrap performance optimism estimation, applied to DCA.}
\usage{
boot_optimismDCA(data, outcome = "newAud", B = 200, thresholds = NULL)
}
\arguments{
\item{data}{Dataset which provides the required columns (see \strong{Details}).}

\item{outcome}{Character string. Name of the outcome variable, in this study it was 'firstAud' (first onset of alcohol use disorder).}

\item{B}{number of bootstrap resamples to run (we set B = 500).}

\item{thresholds}{Vector with numeric values. Selected range of reasonable threshold probabilities, we used: .01, .02, .03, .04, and .05.}
}
\value{
a data.frame with three rows (names: Apparent, Optimism, and Corrected) and as many columns as threshold probabilities were selected:
\enumerate{
\item Apparent Mean apparent net benefit, given the specifiied threshold probability, across all 500 bootstrap versions of the original data.
\item Optimism Mean net benefit difference between the net benefit of the apparent and the cross-validated model performance, for the specified threshold probability.
\item Corrected Apparent net benefit minus optimistic net benefit.
}
}
\description{
Apply the bootstrap evaluation of the model DCA performance.
}
\details{
Dataset must contain the outcome and all eight predictor variables. This function is a modification and extension of the function \code{boot_optimism} of the R package \code{pminternal}.
Marcel Miché modified and extended the original code. Enter predictAUDPsyCoLausDev::boot_optimismDCA in the R console to see the raw code, including the code that I outcommented.
}
\examples{
# See the accompanying R script and this package's vignette,
# section Optimism corrected net benefit.
}
\references{
\insertRef{pbapply2023}{predictAUDPsyCoLaus}

\insertRef{R2024}{predictAUDPsyCoLaus}
}
\author{
Stephen Rhodes (see \strong{Details})
}
