% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_stability1.R
\name{classification_stability1}
\alias{classification_stability1}
\title{Prepare classification stability plot data.}
\usage{
classification_stability1(x, threshold)
}
\arguments{
\item{x}{Output of function \code{validate} from the \code{pminternal} package.}

\item{threshold}{Numeric value between 0 and 1. Estimated risks above the threshold get a predicted 'class' of 1, otherwise 0; see Argument \code{threshold} in \code{pminternal::classification_stability}.}
}
\value{
a data.frame with as many rows as the original data and two columns:
\enumerate{
\item p_orig Apparent predicted probability of each individual, based on the original data.
\item cii Classificiation instability index, i.e., given a specified probability threshold, how often (in percent) has each individual received the opposite classification, compared to \code{p_orig}?
}
A perfect prediction model consistently classifies each individual into one class, whereas switching between classes (for each individual across all bootstrap repetitions) indicates classification instability. Note that the classification, which is based on the original data, is used as the ground truth. That is, the apparent predicted probabilities will always produce misclassifications as well, because no prediction model will ever be deterministically perfect.
See also Riley and Collins, 2023, section 5.4.
}
\description{
Extract data for classification stability.
}
\details{
This function is a modification of the function \code{classification_stability} of the R package \code{pminternal}. The reason to modify it, was that I wanted to visualize the results with ggplot2, not with the R graphics package.
}
\examples{
# See the accompanying R script and this package's vignette,
# section Classification stability.
}
\references{
\insertRef{pminternal2024}{predictAUDPsyCoLaus}
}
\author{
Stephen Rhodes (see \strong{Details})
}
