% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_fun.R
\name{dca_fun}
\alias{dca_fun}
\title{Helper function inside of boot_optimismDCA.}
\usage{
dca_fun(y = NULL, p = NULL, thresholds = NULL)
}
\arguments{
\item{y}{Vector of integer values that are either 0 (= outcome not observed) or 1 (= outcome observed).}

\item{p}{Vector of risk score values (see Arguments of \code{ClinicalUtilityRecal::snbRecalPlot}).}

\item{thresholds}{Vector with numeric values. Selected range of reasonable threshold probabilities, we used: .01, .02, .03, .04, and .05.}
}
\description{
Function to be used in function boot_optimismDCA.
}
\examples{
# This function is used within the function boot_optimismDCA.
}
\references{
\insertRef{mysml2024}{predictAUDPsyCoLaus}

\insertRef{vickers2006decision}{predictAUDPsyCoLaus}
}
\author{
Marcel Miché
}
