% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcurve_stability1.R
\name{dcurve_stability1}
\alias{dcurve_stability1}
\title{Prepare decision curve stability plot data.}
\usage{
dcurve_stability1(x, thresholds)
}
\arguments{
\item{x}{Output of function \code{validate} from the \code{pminternal} package.}

\item{thresholds}{Vector with numeric values. Selected range of reasonable threshold probabilities, we used: .01, .02, .03, .04, and .05.}
}
\value{
a list with b + 1 elements (b = number of bootstrap repetitions). Each element is a dataframe with as many rows as selected threshold probabilities and three columns:
\enumerate{
\item label Character string, always containing 'Prediction model'.
\item threshold Numeric variable with the selected threshold probabilities.
\item net_benefit Numeric variable with the net benefit results for each threshold probability.
}
The first of the list elements contains the net benefits of the apparent performance of the prediction model. The remaining b (we set b = 500) list elements stem from the bootstrap repetitions.
See also Riley and Collins, 2023, section 5.3.
}
\description{
Extract data for decision curve stability.
}
\details{
This function is a modification of the function \code{dcurve_stability} of the R package \code{pminternal}. The reason to modify it, was that I wanted to visualize the results with ggplot2, not with the R graphics package.
}
\examples{
# See the accompanying R script and this package's vignette,
# section Decision curve stability.
}
\references{
\insertRef{pminternal2024}{predictAUDPsyCoLaus}
}
\author{
Stephen Rhodes (see \strong{Details})
}
