% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fun.R
\name{model_fun}
\alias{model_fun}
\title{Helper function inside of boot_optimism1.}
\usage{
model_fun(data)
}
\arguments{
\item{data}{Dataset which provides the required columns (see \strong{Details}).}
}
\description{
Function to be used in function boot_optimism and in boot_optimismDCA.
}
\details{
Dataset must contain the outcome and all eight predictor variables. I removed the three dots for further function arguments from the original function, because we do not use them.
}
\examples{
# This function has not been modified, except for setting the
# outcome to 'newAud' (new onset alcohol use disorder)
}
\author{
Stephen Rhodes
}
