% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDCA.R
\name{plotDCA}
\alias{plotDCA}
\title{Plot DCA results.}
\usage{
plotDCA(allDCA = NULL, bothModels = FALSE)
}
\arguments{
\item{allDCA}{The output 'allDCA' from function \code{visualizeMultipleDCA}.}

\item{bothModels}{Boolean value. TRUE, if logistic regression and random forest shall both be shown in the plot, FALSE, if only logistic regression.}
}
\value{
a list with two elements:
\enumerate{
\item dcaPlot1 Decision curve analysis plot, showing only the mean net benefit curve.
\item dcaPlot2 In addition to dcaPlot1, this plot also shows the interval across all cross-validated net benefit curves.
}
There are three options of intervals for dcaPlot2: mnci = mean and lower to upper 95 percent CI, mnminmax = mean and minimum to maximum, mnq1q3 = mean and first to third quantile.
}
\description{
Function that outputs plotted decision curve analysis results.
}
\examples{
# See the accompanying R script and this package's vignette,
# section 1. Decision curve analysis (DCA), performance measure:
# Net benefit (NB).
}
\references{
\insertRef{ggplot22016}{predictAUDPsyCoLaus}
}
\author{
Marcel Miché
}
