% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_fun.R
\name{pred_fun}
\alias{pred_fun}
\title{Helper function inside of boot_optimism1.}
\usage{
pred_fun(model, data)
}
\arguments{
\item{model}{Output of a fitted logistic regression model.}

\item{data}{Dataset which provides the required columns (see \strong{Details}).}
}
\description{
Function to be used in function boot_optimism and in boot_optimismDCA.
}
\details{
Dataset must contain the outcome and all eight predictor variables. I removed the three dots for further function arguments from the original function, because we do not use them.
}
\examples{
# This function has not been modified. It is used in boot_optimism1
# and in boot_optimismDCA.
}
\author{
Stephen Rhodes
}
