% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictAUDPsyCoLaus.R
\docType{package}
\name{predictAUDPsyCoLaus-package}
\alias{predictAUDPsyCoLaus-package}
\title{Documentation of this predictAUDPsyCoLaus package.}
\description{
Name of this R package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mmiche/predictAUDPsyCoLaus}
  \item Report bugs at \url{https://github.com/mmiche/predictAUDPsyCoLaus/issues}
}

}
