% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictAUDPsyCoLaus.R
\name{predictAUDPsyCoLaus}
\alias{predictAUDPsyCoLaus}
\title{Documentation of this predictAUDPsyCoLaus package.}
\description{
Name of this R package.

All R packages that have been used in this predictAUDPsyCoLaus package, are listed below (see References).

Finally, R packages that were used to develop and publish this predictAUDPsyCoLaus R package on GitHub are: devtools, usethis, roxygen2, and Rdpack. Thanks to the many R package developers!
}
