% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalPotential.R
\name{recalPotential}
\alias{recalPotential}
\title{Recalibration potential.}
\usage{
recalPotential(data = NULL)
}
\arguments{
\item{data}{Dataset which provides the required columns (see \strong{Details}).}
}
\value{
a numeric vector with the two values 0 (= no recalibration method required) and 1 (= some recalibration method is indicated).
}
\description{
Compute recalibration potential.
}
\details{
Dataset must contain the columns 'snbOrig', 'snbStdRecal', 'snbConstRecal', and 'oneSEfromMax'; this dataset is returned by the function \code{snbVals}.
}
\examples{
# See the accompanying R script and this package's vignette,
# section Recalibration: Potential for improved net benefit?
}
\author{
Marcel Miché
}
