% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snbRsmpPlot.R
\name{snbRsmpPlot}
\alias{snbRsmpPlot}
\title{Plot recalibration data.}
\usage{
snbRsmpPlot(data = NULL)
}
\arguments{
\item{data}{Dataset which provides the required columns (see \strong{Details}).}
}
\value{
a ggplot2 visualization with three facet windows
\enumerate{
\item facet 1 shows 
\item facet 2 shows
\item facet 3 
}
}
\description{
Plot the extracted standardized net benefit results.
}
\details{
Dataset must contain the columns 'snbOrig', 'snbStdRecal', 'snbConstRecal', and 'oneSEfromMax'; this dataset is returned by the function \code{snbVals}.
}
\examples{
# This function is used by its wrapper function \code{snbVals}.
}
\references{
\insertRef{tidyr2024}{predictAUDPsyCoLaus}

\insertRef{ggplot22016}{predictAUDPsyCoLaus}
}
\author{
Marcel Miché
}
