% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snbVals.R
\name{snbVals}
\alias{snbVals}
\title{Repeatedly apply snbRecalPlotNum.}
\usage{
snbVals(y = NULL, p = NULL, r = NULL)
}
\arguments{
\item{y}{Vector of integer values that are either 0 (= outcome not observed) or 1 (= outcome observed).}

\item{p}{Vector of risk score values (see Arguments of \code{ClinicalUtilityRecal::snbRecalPlot}).}

\item{r}{Numeric value. Clinically relevant risk threshold (see Arguments of \code{ClinicalUtilityRecal::snbRecalPlot}).}
}
\value{
Output which is returned by the function \code{snbRecalPlotNum}. In other words, \code{snbVals} is a wrapper function.
Note: In a proper R package, a wrapper function remains invisible (a so-called helper function). However, this R package is part of the supplementary material of a publication, which is why we made each function visible.
}
\description{
Function that enables repeated application of snbRecalPlotNum.
}
\examples{
# See the accompanying R script and this package's vignette,
# section Recalibration: Potential for improved net benefit?
}
\references{
\insertRef{ClinicalUtilityRecal2020}{predictAUDPsyCoLaus}
}
\author{
Marcel Miché
}
