% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeMultipleDCA.R
\name{visualizeMultipleDCA}
\alias{visualizeMultipleDCA}
\title{Prepare data for plotDCA.}
\usage{
visualizeMultipleDCA(dcaLs = NULL, dcaSelectType = NULL)
}
\arguments{
\item{dcaLs}{The output dcaLs from function \code{computeRelevantResults} from the \code{mysml} package.}

\item{dcaSelectType}{Character string. One out of these three options (see \strong{Details}): mnci, mnminmax, mnq1q3.}
}
\value{
a list with three elements:
\enumerate{
\item allDCA A dataframe needed by the function \code{plotDCA}.
\item logregNB A dataframe with as many rows as selected threshold probabilities. The columns contain summaries of the net benefit, i.e., the median, the mean, the minimum, the first and the third quantile, and the maximum.
\item rfNB Same as logregNB, only for the random forest model.
}
}
\description{
Function that prepares the data for the function plotDCA.
}
\examples{
# See the accompanying R script and this package's vignette,
# section 1. Decision curve analysis (DCA), performance measure:
# Net benefit (NB).
}
\references{
\insertRef{mysml2024}{predictAUDPsyCoLaus}

\insertRef{dplyr2023}{predictAUDPsyCoLaus}

\insertRef{magrittr2022}{predictAUDPsyCoLaus}
}
\author{
Marcel Miché
}
