% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictAUDPsyCoLaus.R
\docType{package}
\name{predictAUDPsyCoLaus-package}
\alias{predictAUDPsyCoLaus-package}
\title{predictAUDPsyCoLaus: predictAUDPsyCoLaus}
\description{
Provide full transparency of analysis code for a clinical prediction model of new onset alcohol use disorder.
}
\references{
\insertRef{R2024}{predictAUDPsyCoLaus}

\insertRef{devtools2021}{predictAUDPsyCoLaus}

\insertRef{usethis2021}{predictAUDPsyCoLaus}

\insertRef{roxygen2}{predictAUDPsyCoLaus}

\insertRef{Rdpack2022}{predictAUDPsyCoLaus}

\insertRef{dplyr2023}{predictAUDPsyCoLaus}

\insertRef{magrittr2022}{predictAUDPsyCoLaus}

\insertRef{tidyr2024}{predictAUDPsyCoLaus}

\insertRef{ranger2017}{predictAUDPsyCoLaus}

\insertRef{mysml2024}{predictAUDPsyCoLaus}

\insertRef{ggplot22016}{predictAUDPsyCoLaus}

\insertRef{cowplot2024}{predictAUDPsyCoLaus}

\insertRef{gridExtra2017}{predictAUDPsyCoLaus}

\insertRef{pmcalibration2025}{predictAUDPsyCoLaus}

\insertRef{pminternal2025}{predictAUDPsyCoLaus}

\insertRef{ClinicalUtilityRecal2020}{predictAUDPsyCoLaus}

\insertRef{rms2025}{predictAUDPsyCoLaus}

\insertRef{precrec2017}{predictAUDPsyCoLaus}

\insertRef{samplesizedev2024}{predictAUDPsyCoLaus}

\insertRef{modgo2022}{predictAUDPsyCoLaus}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mmiche/predictAUDPsyCoLaus}
  \item Report bugs at \url{https://github.com/mmiche/predictAUDPsyCoLaus/issues}
}

}
